refactor: change Int64 for frame num to Int32

+++ src/fileio.jl
@@ -53,7 +53,7 @@ function readdcd(filename::String; frame_indices::Union{Vector, OrdinalRange, Co
         # read body block
         header_size = position(io)
         coordblocksize = (8 + 4 * number_of_atom) * 3
-        total_frame = Int64((file_size - header_size) / coordblocksize)
+        total_frame = Int32((file_size - header_size) / coordblocksize)
         if typeof(target_frame_indices) == Colon
             target_frame_indices = 1:total_frame
         end
